change distance result brief information to intptr_t
The distance result brief information stored the brief information
in an `int`. However, some primatives can only be stored by pointer
(namely octree nodes), and `int` is not wide enough on 64-bit
platforms to safely store a pointer. So upgrade the brief
information to be `intptr_t` which is guaranteed to be wide enough.

+++ include/fcl/narrowphase/distance_result.h
@@ -78,13 +78,13 @@ public:
   /// if object 1 is mesh or point cloud, it is the triangle or point id
   /// if object 1 is geometry shape, it is NONE (-1),
   /// if object 1 is octree, it is the id of the cell
-  int b1;
+  intptr_t b1;
 
   /// @brief information about the nearest point in object 2
   /// if object 2 is mesh or point cloud, it is the triangle or point id
   /// if object 2 is geometry shape, it is NONE (-1),
   /// if object 2 is octree, it is the id of the cell
-  int b2;
+  intptr_t b2;
 
   /// @brief invalid contact primitive information
   static const int NONE = -1;
